+++ CHANGELOG.markdown
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2023-04-12
+
+### Fixed
+
+*  Fixed incorrectly mergiing instructions across `printflush()` calls ([#91](https://github.com/francois/mindcode/issues/91)).
+
 ## 2023-04-08
 
 ### Added